Organize device modules into package
