+++ README.md
@@ -6,10 +6,15 @@ cell through an AL2205 IO-Link hub connected to an AL1342 IO master.
 ## Contents
 
 ```
-IO_master.py               # Modbus communication with the AL1342
-AL2205_Hub.py              # Access analog values from the AL2205 hub
-LoadCell_LCM300.py         # Convert analog values to force readings
-test_read_five_load_cells.py # Command line utility to read up to five load cells
+devices/                   # Hardware interface modules
+  IO_master.py             # Modbus communication with the AL1342
+  AL2205_Hub.py            # Access analog values from the AL2205 hub
+  LoadCell_LCM300.py       # Convert analog values to force readings
+  PressureSensor_PQ3834.py # PQ3834 pressure sensor helper
+  FlowPressure_SD9500.py   # SD9500 flow/pressure sensor helper
+  FlowSensor_SD6020.py     # SD6020 flow sensor helper
+Iris_EOL_Fixture.py        # Convenience functions for common tasks
+Sample_Test_Script.py      # Example script exercising the fixture
 ```
 
 ## Installation
@@ -26,9 +31,9 @@ example below assumes the AL2205 hub is connected to port 1 and the load cell
 is on channel X1.0.
 
 ```python
-from IO_master import IO_master
-from AL2205_Hub import AL2205Hub
-from LoadCell_LCM300 import LoadCellLCM300
+from devices.IO_master import IO_master
+from devices.AL2205_Hub import AL2205Hub
+from devices.LoadCell_LCM300 import LoadCellLCM300
 
 io = IO_master()                          # Defaults to IP 192.168.100.1
 hub = AL2205Hub(io, port_number=1)        # Hub connected to port 1
@@ -36,12 +41,6 @@ cell = LoadCellLCM300(hub, x1_index=0)    # Load cell on channel X1.0
 
 print(cell.read_force())                  # Read force in newtons (default)
 print(cell.read_force("lbf"))             # Read force in pounds-force
-```
-
-To read up to five load cells from the command line:
-
-```
-python test_read_five_load_cells.py 192.168.100.1
 ```
 
 ## License
